@@ -1,22 +1,21 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
-import reportWebVitals from './reportWebVitals';
-import App from './app/App';
-import {Provider} from 'react-redux';
-import {store} from './app/store';
-import {BrowserRouter} from 'react-router-dom';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
+import {createRoot} from 'react-dom/client';
+import {Provider} from "react-redux";
+import {store} from "./store/store";
 
-ReactDOM.render(
+
+
+
+
+const container  = document.getElementById('root') as HTMLElement
+const root = createRoot(container);
+root.render(
     <Provider store={store}>
-        <BrowserRouter>
-            <App/>
-        </BrowserRouter>
-    </Provider>,
-    document.getElementById('root')
+        <App/>
+    </Provider>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+serviceWorker.unregister();
